+++ src/main/java/Ui.java
@@ -1,0 +1,7 @@
+public class Ui {
+
+    public void showLoadingError(){
+        System.out.println("Error while loading file");
+    }
+
+}
